Level Loading part1
Added an S-INIT Packet which passes the server's mapping of entities (render object instances for the client) to the client, with their parent name, that can be loaded using its model and textures on the client. Next to add is actual client interpretation of this information (i.e. displaying level) and updating it when it changes, as well as sending models and textures to the client for server by server resources.